Change the type of radii from int to double
Floating numbers should be used to accommodate the changes that will be
done in the future.

+++ src/main/java/com/targus/ui/Sensor.java
@@ -45,7 +45,7 @@ public class Sensor extends Circle {
         return communicationRangeRadius;
     }
 
-    public static void setRadii(int communicationRangeRadius, int sensingRangeRadius){
+    public static void setRadii(double communicationRangeRadius, double sensingRangeRadius){
         Sensor.communicationRangeRadius = communicationRangeRadius;
         Sensor.sensingRangeRadius = sensingRangeRadius;
     }
